+++ app/Main.hs
@@ -1,28 +1,22 @@
 module Main (main) where
 
+import GHC.Conc (setUncaughtExceptionHandler)
 import Shrun.Configuration.Env (makeEnvAndShrun)
 import Shrun.Prelude hiding (IO)
 import Prelude (IO)
 
 main :: IO ()
 main = do
-  -- NOTE: Seems redundant, but sadly it isn't.
-  --
-  -- doNothingOnSuccess is solely used to stop ExitSuccess from poisoning the
-  -- error code because GHC determines exit success/failure based on the
-  -- presence of an uncaught exception. However, we still need to rethrow
-  -- the failure.
-  --
-  -- Moreover, we don't really want to print a CallStack for ExitFailure,
-  -- as we are throwing that whenever a command fails, and the CallStack
-  -- is just unhelpful noise.
-  setUncaughtExceptionHandlerDisplay
+  setUncaughtExceptionHandler handleEx
 
-  makeEnvAndShrun @IO @ConsoleRegion `catch` doNothingOnSuccess
+  makeEnvAndShrun @IO @ConsoleRegion
   where
-    -- We need to catch ExitCode so that optparse applicative's --help
-    -- does not set the error code to failure...but then we need to rethrow
-    -- failures.
-    doNothingOnSuccess :: ExitCode -> IO ()
-    doNothingOnSuccess ExitSuccess = pure ()
-    doNothingOnSuccess ex@(ExitFailure _) = throwM ex
+    handleEx ex = case fromException ex of
+      -- Do not print ExitCode
+      Just ExitSuccess -> pure ()
+      Just (ExitFailure _) -> pure ()
+      Nothing -> case fromException ex of
+        -- Do not print term exception, since we handle it elsewhere. It
+        -- reaches this point purely to kill the program.
+        Just MkTermException -> pure ()
+        Nothing -> putStrLn $ displayException ex