+++ src/Shrun/ShellT.hs
@@ -7,6 +7,24 @@ module Shrun.ShellT
   )
 where
 
+import Effects.System.Posix.Signals
+  ( MonadPosixSignals
+      ( awaitSignal,
+        blockSignals,
+        getPendingSignals,
+        getSignalMask,
+        installHandler,
+        queryStoppedChildFlag,
+        raiseSignal,
+        scheduleAlarm,
+        setSignalMask,
+        setStoppedChildFlag,
+        signalProcess,
+        signalProcessGroup,
+        unblockSignals
+      ),
+  )
+import Effects.System.Posix.Signals qualified as Signals
 import Shrun.Configuration.Data.Notify.System
   ( NotifySystemP (AppleScript, DBus, NotifySend),
   )
@@ -37,7 +55,7 @@ newtype ShellT env m a = MkShellT (ReaderT env m a)
       MonadIORef,
       MonadMask,
       MonadPathWriter,
-      MonadTypedProcess,
+      MonadProcess,
       MonadReader env,
       MonadSTM,
       MonadThread,
@@ -45,6 +63,9 @@ newtype ShellT env m a = MkShellT (ReaderT env m a)
       MonadThrow
     )
     via (ReaderT env m)
+
+unShellT :: ShellT env m a -> ReaderT env m a
+unShellT (MkShellT rdr) = rdr
 
 -- | Runs a 'ShellT' with the given @env@.
 runShellT :: forall m env a. ShellT env m a -> env -> m a
@@ -61,7 +82,7 @@ deriving newtype instance (MonadRegionLogger m) => MonadRegionLogger (ShellT (En
 
 instance
   ( MonadDBus m,
-    MonadTypedProcess m
+    MonadProcess m
   ) =>
   MonadNotify (ShellT (Env r) m)
   where
@@ -73,3 +94,49 @@ instance
       sendNote (DBus client) = DBus.notifyDBus client note
       sendNote NotifySend = NotifySend.notifyNotifySend note
       sendNote AppleScript = AppleScript.notifyAppleScript note
+
+-- REVIEW: Would be nice if we could derive this...
+
+instance (MonadPosixSignals m) => MonadPosixSignals (ShellT env m) where
+  raiseSignal = MkShellT . raiseSignal
+  {-# INLINEABLE raiseSignal #-}
+
+  signalProcess s = MkShellT . signalProcess s
+  {-# INLINEABLE signalProcess #-}
+
+  signalProcessGroup s = MkShellT . signalProcessGroup s
+  {-# INLINEABLE signalProcessGroup #-}
+
+  installHandler s h m = MkShellT $ do
+    hFromM <$> installHandler s (hToM h) m
+    where
+      hFromM = Signals.mapHandler MkShellT
+      hToM = Signals.mapHandler unShellT
+  {-# INLINEABLE installHandler #-}
+
+  getSignalMask = MkShellT getSignalMask
+  {-# INLINEABLE getSignalMask #-}
+
+  setSignalMask = MkShellT . setSignalMask
+  {-# INLINEABLE setSignalMask #-}
+
+  blockSignals = MkShellT . blockSignals
+  {-# INLINEABLE blockSignals #-}
+
+  unblockSignals = MkShellT . unblockSignals
+  {-# INLINEABLE unblockSignals #-}
+
+  scheduleAlarm = MkShellT . scheduleAlarm
+  {-# INLINEABLE scheduleAlarm #-}
+
+  getPendingSignals = MkShellT getPendingSignals
+  {-# INLINEABLE getPendingSignals #-}
+
+  awaitSignal = MkShellT . awaitSignal
+  {-# INLINEABLE awaitSignal #-}
+
+  setStoppedChildFlag = MkShellT . setStoppedChildFlag
+  {-# INLINEABLE setStoppedChildFlag #-}
+
+  queryStoppedChildFlag = MkShellT queryStoppedChildFlag
+  {-# INLINEABLE queryStoppedChildFlag #-}